pyproject.toml
[tool.poetry]
name = "momo"
version = "0.1.0"
description = ""
authors = ["sliverKi <lovee2756@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "~4.0"
djangorestframework = "^3.14.0"
dj-rest-auth = "^3.0.0"
django-cors-headers = "^3.14.0"
django-multiupload = "^0.6.1"
django-allauth = "^0.54.0"
django-rest-auth = "^0.9.5"
django-environ = "^0.10.0"
drf-yasg = "^1.21.5"
django-elasticsearch-dsl = "^7.3"
django-elasticsearch-dsl-drf = "^0.22.5"
django-seed = "^0.3.1"
faker = "^18.9.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.6"
elasticsearch = "7.14.0"
python-logstash = "^0.4.8"
json-log-formatter = "^0.5.2"
django-debug-toolbar = "^4.1.0"
gunicorn = "^20.1.0"
whitenoise = "^6.4.0"
dj-database-url = "^2.0.0"
sentry-sdk = "^1.24.0"
boto3 = "^1.28.37"
django-storages = "^1.13.2"
django-redis = "^5.4.0"
trio-websocket = "^0.11.1"
ip2geotools = "^0.1.6"
geoip2 = "^4.8.0"
mysqlclient = "^2.2.1"
django-silk = "^5.1.0"
requests-aws4auth = "^1.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
